+++ order-service/src/main/java/com/anurag/order/domain/FoodMenu.java
@@ -1,5 +1,7 @@
 package com.anurag.order.domain;
 
+import java.math.BigDecimal;
+
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
@@ -13,7 +15,7 @@ public class FoodMenu {
 	private String foodName;
 	private String foodDescription;
 	private String foodCatalog;
-	private double foodPrice;
+	private BigDecimal foodPrice;
 	private Long restaurantId;
 
 }
